Removes DiffableLongIdItem And Uses Table Schema Instead
Summary:
Previously it was using a special interface, DiffableLongIdItem,
in order to get the ids by which items for animation are compared.
This removes this interface requirement and simply uses the primary
key of the item in order to retrieve a unique id that can be used
for the row diffing.